Add pretty much literal translation of wiki pseudocode
There were some hairy spots that weren't clear.

One interesting idea was that the pseudocode presumed the ability to
have an initial run of a while loop without checking the condition. It
used this to set a variable that was set in the condition.

What's the correct style? Something to ask on Zulip I guess.